backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
pyjwt>=2.8.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
